+++ Engine/renderer/graphics_pipeline.h
@@ -18,7 +18,7 @@ namespace engine
   struct fgraphics_pipeline
   {
     void reserve_parameters(uint32_t num);
-    void add_constant_parameter(uint32_t index, uint32_t shader_register, uint32_t register_space, size_t size, D3D12_SHADER_VISIBILITY visibility);
+    void add_constant_parameter(uint32_t index, uint32_t shader_register, uint32_t register_space, uint32_t size, D3D12_SHADER_VISIBILITY visibility);
     void add_shader_respurce_view_parameter(uint32_t index, uint32_t shader_register, uint32_t register_space, D3D12_ROOT_DESCRIPTOR_FLAGS flags, D3D12_SHADER_VISIBILITY visibility);
     void add_constant_buffer_view_parameter(uint32_t index, uint32_t shader_register, uint32_t register_space, D3D12_ROOT_DESCRIPTOR_FLAGS flags, D3D12_SHADER_VISIBILITY visibility);
     void add_unordered_access_view_parameter(uint32_t index, uint32_t shader_register, uint32_t register_space, D3D12_ROOT_DESCRIPTOR_FLAGS flags, D3D12_SHADER_VISIBILITY visibility);
